Apply OCEAMIC Laayoune II branding and fix Phase 6 QMS workflow issues
Branding: official logo/wordmark in the sidebar (expanded and mobile), login
page and favicon; brand palette (dark blue primary, cyan accent, extracted
from the logo) replaces the generic green for primary actions and the active
nav state, green now reserved for semantic success only. The heavy
near-black KPI card is now brand blue.

QMS audit fixes, backed by domain/service-layer enforcement (not just UI):
- Non-conformity status transitions are now constrained
  (NONCONFORMITY_ALLOWED_TRANSITIONS) instead of exposing all six statuses
  as flat buttons; a fresh record can no longer jump straight to CLOTUREE.
- A CRITIQUE severity with a BASSE/NORMALE priority now requires explicit
  confirmation instead of silently coexisting; NCRs auto-created from an
  audit finding or complaint now derive priority from severity instead of
  always defaulting to NORMALE.
- CAPA/CAPA action forms gained a due-date field, required for
  HAUTE/URGENTE priority, so the "CAPA en retard" KPI can actually fire.
- CAPA effectiveness now reads "Efficace"/"Non efficace" instead of the
  ambiguous "Conforme"/"Non conforme".
- A completed audit with open findings now says so explicitly instead of
  implying every follow-up is closed; fixed a real bug where
  audit_progress's finding counts were returned as strings instead of
  numbers (missing ::integer cast).
- "Accès rapides" (navigation) and "Actions rapides" (create shortcuts)
  are now visually and functionally distinct on the QMS home page.
- Verified end-to-end that entity links (e.g. complaint → shipment →
  pallet → Lot PF → cycle → run → Lot MP) are genuine relational joins,
  and that all 105 Phase 6 enum values resolve to a French label.

Seed data updated to demonstrate realistic combinations (a genuinely
overdue CAPA/action, correct severity/priority pairing) instead of always
using null due dates. Tests added for each corrected behavior; full suite
(126 tests) and both builds pass.

+++ oceamic-ims/server/src/db/migrations/025_qms_views.sql
@@ -48,12 +48,16 @@ SELECT c.id AS capa_id,
 -- Audit progress: responses recorded against the checklist attached to the
 -- audit (if any), and open findings - the "Constats / Actions ouvertes"
 -- columns of the Audit screen (section 42).
+-- Every count is cast to ::integer, the same discipline as
+-- capa_action_progress above: left uncast, COUNT(*) is bigint and the pg
+-- driver returns it as a string, silently breaking any numeric comparison
+-- (`=== 3`) on the API response.
 CREATE VIEW audit_progress AS
 SELECT a.id AS audit_id,
-       COALESCE(items.total_items, 0)      AS total_checklist_items,
-       COALESCE(responses.response_count, 0) AS response_count,
-       COALESCE(findings.finding_count, 0)   AS finding_count,
-       COALESCE(findings.open_finding_count, 0) AS open_finding_count
+       COALESCE(items.total_items, 0)::integer      AS total_checklist_items,
+       COALESCE(responses.response_count, 0)::integer AS response_count,
+       COALESCE(findings.finding_count, 0)::integer   AS finding_count,
+       COALESCE(findings.open_finding_count, 0)::integer AS open_finding_count
   FROM audits a
   LEFT JOIN LATERAL (
         SELECT COUNT(*) AS total_items FROM audit_checklist_items i
